Move all azure api into lib/cloud/azure
