+++ lib/cloud/azure/constants.go
