+++ lib/cloud/azure/interfaces.go
